scripts for preparing training data
